pyproject.toml
[tool.poetry]
name = "git-gut"
version = "0.1.0"
description = ""
authors = ["mogryo <tad.baikstis@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.1.7"
gitdb = "^4.0.11"
GitPython = "^3.1.43"
smmap = "^5.0.1"
prettytable = "^3.10.0"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
pylint = "^3.2.3"
SQLAlchemy = "^2.0.30"
typing-extensions = "^4.12.1"
pydash = "^8.0.3"
rich = "^13.9.3"


[tool.poetry.group.dev.dependencies]
black = {extras = ["d"], version = "^24.4.2"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
venvPath = "/Users/tadas.baikstis/Library/Caches/pypoetry/virtualenvs/git-gut-YknkvPg0-py3.12"
venv = "/Users/tadas.baikstis/Library/Caches/pypoetry/virtualenvs/git-gut-YknkvPg0-py3.12/bin/python"
